add some translations of accentuated characters (patch submitted by James Van Zandt to debian)
Add Unicode translations for all the accents on p. 52 of the TeXbook
using pre-composed characters where available, otherwise characters
with combining class "above" or "below".

Add Unicode translations for all the math accents on p. 135 of the
TeXbook similarly, but using <em>...</em> to get italic characters.

Correct the translations for \~u and \~U.  (They had the hex
Unicode values, but lacked the "x".)

Correct the translation of \dot.  (It was translated as a common
period, but should be an accent.)

Correct the translation for \models to U+22a7, as per
https://www.compart.com/en/unicode/U+22A8, (The former translation
U+22a8 is the symbol for "true", which is similar to the one for
"models" but not as tall.  There's no record of a symbol "not
models".  U+22ad (decimal 8877) is the Unicode character for "not
true", which is the symbol for "true" with a stroke added.  However,
neither \true nor \false are standard macros in LaTeX, so there is no
need to provide those translations.)

Add translations for dotless i and j without accents, \div,
\ddagger, \LongLeftArrow, and many more macros - all the plain TeX
symbols in the document "Every symbol (most symbols) defined by
unicode-math" by Will Robertson, v0.8m of 2018-07-29, found here:
http://mirrors.ibiblio.org/CTAN/macros/latex/contrib/unicode-math/unimath-symbols.pdf